fix(SFrameGrabber+SMatricesReader): current index position in setStep
Calling setStep slots changed the current step but not the current index position. We need to also update this indices position because it's current value is update right after reading a next image/matrix. If this indices are not changed setStep slots, the step will be take into account only on the next call of readNext slots.

+++ Bundles/io/ioTimeline/include/ioTimeline/SMatricesReader.hpp
@@ -127,7 +127,7 @@ private:
     void pause();
 
     /// SLOT: Set step used on readPrevious/readNext slots
-    void setStep(unsigned int _step);
+    void setStep(int _step);
 
     /// Read matrices (this function is set to the worker)
     void readMatrices();
@@ -154,7 +154,7 @@ private:
     bool m_useTimelapse;
 
     /// Step between two matrices when calling readNext() and readPrevious() slots
-    unsigned int m_step;
+    unsigned long m_step;
 };
 
 } // ioTimeline
